+++ modules/narrativepipeline/NarrativePipeline.py
@@ -8,18 +8,19 @@ from transformers import AdamW
 
 
 from modules.narrativepipeline.utils_origin import CustomDataset, build_vocab_PGD, Vocab
-from modules.pg_decoder.PointerGeneratorDecoder import PointerGeneratorDecoder
-from modules.utils import EmbeddingLayer, check_exist, transpose, get_scores
+from modules.ans_infer.PointerGeneratorDecoder import PointerGeneratorDecoder
+from modules.utils import check_exist, transpose, get_scores
 from modules.Reasoning.IAL import IntrospectiveAlignmentLayer
+from modules.finegrained.Embedding import EmbeddingLayer
 from configs import args, logging, PATH
 
 class  NarrativePipeline(torch_nn.Module):
-    def __init__(self):
+    def __init__(self, vocab):
         super().__init__()
 
         self.embd_layer = EmbeddingLayer()
         self.reasoning  = IntrospectiveAlignmentLayer()
-        self.pg_decoder = PointerGeneratorDecoder()
+        self.ans_infer  = PointerGeneratorDecoder(vocab)
 
     def forward(self, ques, ques_len, contx, contx_len, ans, ans_len, ans_mask):
         # ques      : [b, seq_len_ques, d_embd]
@@ -51,7 +52,7 @@ class  NarrativePipeline(torch_nn.Module):
         ####################
         # Generate answer with PGD
         ####################
-        pred    = self.pg_decoder(Y, H_q, ans, ans_len, ans_mask)
+        pred    = self.ans_infer(Y, H_q, ans, ans_len, ans_mask)
         # pred: [batch, max_len_ans, d_vocab + seq_len_cntx]
 
         return pred
@@ -212,7 +213,7 @@ class Trainer():
         ###############################
         # Defind model and associated stuffs
         ###############################
-        model       = NarrativePipeline().to(args.device)
+        model       = NarrativePipeline(self.vocab).to(args.device)
         optimizer   = AdamW(model.parameters(), lr=args.lr, weight_decay=args.w_decay)
         criterion   = torch_nn.CrossEntropyLoss()
 
@@ -260,11 +261,11 @@ class Trainer():
             pred (tensor): predicted tensor
             ans_tok_idx (tensor): target answer
         """
-        # pred: [batch, d_vocab + seq_len_cntx, max_len_ans]
-        # ans_tok_idx: [batch, max_len_ans]
+        # pred: [batch, d_vocab + seq_len_cntx, seq_len_ans]
+        # ans_tok_idx: [batch, seq_len_ans]
 
         pred_   = torch.argmax(torch_f.log_softmax(pred, 1), 1)
-        # pred_: [batch, max_len_ans]
+        # pred_: [batch, seq_len_ans]
 
         bleu_1, bleu_4, meteor, rouge_l = 0, 0, 0, 0
 
@@ -293,7 +294,7 @@ class Trainer():
         ###############################
         # Defind model and associated stuffs
         ###############################
-        model       = NarrativePipeline().to(args.device)
+        model       = NarrativePipeline(self.vocab).to(args.device)
         criterion   = torch_nn.CrossEntropyLoss()
         model       = model.to(args.device)